metadataextractorsearch/requirements.txt
django
djangorestframework
django-cors-headers
cryptography
pillow
django-filter